Added measurements
Added measurements to display for speed/heading/altitude. Moved aircraft category mapping to the static file for consistency.
