+++ sage-radar/src/components/adsb-map/adsb-details.js
@@ -1,30 +1,6 @@
 import React from 'react';
+import { categoryAircraftMap } from "./static.js"
 
-
-/* This component displays the selected Aircrafts ADS-B data in a seperate div from the Leaflet map
- * If no aircraft is selected, it displays a message.
- * If an aircraft is selected, it displays the aircraft's ADS-B data. */
-
-// Array associating the Category codes with their definitions
-const categoryAircraftMap = {
-    "A1": "Light",
-    "A2": "Small",
-    "A3": "Large",
-    "A4": "High Vortex Large",
-    "A5": "Heavy",
-    "A6": "High Performance",
-    "A7": "Rotorcraft",
-    "B1": "Glider/Light Sport",
-    "B2": "Lighter-than-Air",
-    "B3": "Parachutist/Skydiver",
-    "B4": "Ultralight/Hang-Glider",
-    "B6": "UAV",
-    "B7": "Space",
-    "C1": "Surface Emergency",
-    "C2": "Surface Service",
-    "C3": "Surface Obstruction",
-    "Unknown": "Unknown"
-}
 
 
 // Function to generate the component. 
@@ -47,9 +23,9 @@ const SelectedAircraftInfo = ({ selectedAircraft }) => {
                 <>
                     <p><b>Aircraft ID:</b> {selectedAircraft.M.hex.S}</p>
                     <p><b>Callsign:</b> {selectedAircraft.M.flight?.S || 'N/A'}</p>
-                    <p><b>Heading:</b> {selectedAircraft.M.track?.N || 'N/A'}</p>
-                    <p><b>Speed:</b> {selectedAircraft.M.gs?.N || 'N/A'}</p>
-                    <p><b>Altitude:</b> {selectedAircraft.M.alt_baro?.N || 'N/A'}</p>
+                    <p><b>Heading:</b> {selectedAircraft.M.track?.N || 'N/A'}°</p>
+                    <p><b>Speed:</b> {selectedAircraft.M.gs?.N || 'N/A'}kts</p>
+                    <p><b>Altitude:</b> {selectedAircraft.M.alt_baro?.N || 'N/A'}ft</p>
                     <p><b>Mode 3A:</b> {selectedAircraft.M.squawk?.S || 'N/A'}</p>
                     <p><b>Category:</b> {catDesc}</p>
                     <p><b>Latitude:</b> {selectedAircraft.M.lat.N}</p>
